Add Redux action to form onClick
When we type in the form or submit a search,
we fire a redux action that notifies the store
we're fetching data and ResultsView renders
the spinner if fetchingData prop is true.
May delete this feature later because the
api search is super fast and makes rendering
a spinner while fetching data kind of hacky.